filesystem library with his asyn methods
This methods are wrote using async in that case they are sent to
the thread pool. Also this method are controlled by the
event loop. The normal method like console log with message were always controlled by
the call stack.
The null are writted controlled by the event queue. The problem is tah the last print of deleted
is after the resolution of the callback .